stager: when checking if file is already staged, validate etag
If it is a mismatch, download the file from S3 (fast) so we may then
reset the etag while uploading.

Technically it is possible to copy a file to itself within a bucket,
and that might rest the etag if done with the correct chunksize, but
this is more staightforward.